requirements.txt
dj-rest-auth
django
django-allauth
django-filter
djangorestframework
faker
drf_spectacular